@@ -1,180 +1,283 @@
-Shifts::Application.routes.draw do |map|
+Shifts::Application.routes.draw do
 
-  map.resources :shifts_tasks
-
-  map.resources :tasks, :member => {:make_entry => :post}
-
-  map.resources :template_time_slots
-
-  map.resources :stickies
-
-  map.resources :announcements
-
-  map.resources :links
-
-  map.resources :repeating_events
-
-  map.resources :calendars, :member => {:toggle => :post, :prepare_copy => :get, :copy => :post, :apply_schedule => [:get, :post]}, :collection => {:prepare_wipe_range => :get, :wipe_range => :post}
-
-  map.resources :punch_clock_sets
-
-  map.with_options :controller => 'first_run' do |f|
-    f.first_app_config 'firstrun/first_app_config', :action => 'new_app_config', :method => 'get'
-    f.first_department 'firstrun/first_department', :action => 'new_department', :method => 'get'
-    f.first_user 'firstrun/first_user', :action => 'new_user', :method => 'get'
-    f.create_first_app_config 'firstrun/create_first_app_config', :action => 'create_app_config', :method => 'post'
-    f.create_first_department 'firstrun/create_first_department', :action => 'create_department', :method => 'post'
-    f.create_first_user 'firstrun/create_first_user', :action => 'create_user', :method => 'post'
+  resources :tasks do
+    post :make_entry, on: :member
+    get 'update_tasks', on: :collection
   end
 
-  map.login "login", :controller => 'user_sessions', :action => 'new'
-  map.logout "logout", :controller => 'user_sessions', :action => 'destroy'
+  resources :shifts_tasks
+  resources :calendar_feeds
+  resources :template_time_slots
+  resources :stickies
+  resources :announcements
+  resources :links
+  resources :repeating_events
+  resources :restrictions
+  resources :punch_clocks
+  resources :punch_clock_sets
+  resources :user_sessions
+  resources :password_resets
+  resources :user_configs, :only => [:edit, :update]
+  resources :sub_requests
+  resources :notices do
+    get 'archive', on: :collection
+    get 'update_message_center', on: :collection
+  end
+  resources :payform_item_sets
+  resources :payform_sets
+  resources :department_configs, :only => [:edit, :update]
+  resources :requested_shifts
+  resources :templates
+  resources :payform_items
+  resources :requested_shifts
+  resources :templates
 
-  map.resources :app_configs, :only => [:edit, :update]
 
-  map.edit_app_config "app_config", :controller => 'app_configs', :action => 'edit', :method => 'get'
 
-  map.resources :punch_clocks
-  map.resources :restrictions
-  map.email_reminders "email_reminders", :controller => 'payforms', :action => 'email_reminders'
-  map.reminders_advanced_options "reminders_advanced_options", :controller => 'payforms', :action => 'reminders_advanced_options'
-  map.warnings_advanced_options  "warnings_advanced_options", :controller => 'payforms', :action => 'warnings_advanced_options'
+  resources :stats do
+    collection do
+      get  :for_user
+      post :for_user
+      get  :for_location
+      post :for_location
+      get  :index
+      post :index
+    end
+  end
+
+
+  resources :app_configs, :only => [:edit, :update]
+  get "app_config" => 'app_configs#edit', as: :edit_app_config
+
+  resources :calendars do
+    member do
+      get  :prepare_copy
+      get  :apply_schedule
+      post :apply_schedule
+      post :toggle
+      post :copy
+    end
+    collection do
+     get :prepare_wipe_range
+     post :wipe_range
+    end
+  end
+
+
+  scope '/firstrun' do
+    with_options controller: 'first_run' do |f|
+      f.match '/first_app_config', :action => 'new_app_config', :method => 'get', as: :first_app_config
+      f.match '/first_department', :action => 'new_department', :method => 'get', as: :first_department
+      f.match '/first_user', :action => 'new_user', :method => 'get', as: :first_user
+      f.match '/create_first_app_config', :action => 'create_app_config', :method => 'post', as: :create_first_app_config
+      f.match '/create_first_department', :action => 'create_department', :method => 'post', as: :create_first_department
+      f.match '/create_first_user', :action => 'create_user', :method => 'post', as: :create_first_user
+    end
+  end
+
+  match 'login' => 'user_sessions#new', as: :login
+  match 'logout' => 'user_sessions#destroy', as: :logout
+
+  match "email_reminders", :controller => 'payforms', :action => 'email_reminders', as: :email_reminders
+  match "reminders_advanced_options", :controller => 'payforms', :action => 'reminders_advanced_options', as: :reminders_advanced_options
+  match  "warnings_advanced_options", :controller => 'payforms', :action => 'warnings_advanced_options', as: :warnings_advanced_options
 
   #TODO: get rid of sessions controller and move logout action to user_session controller and name it cas_logout
-  map.cas_logout "cas_logout", :controller => 'sessions', :action => 'logout'
-
-   # routes for calendar_feeds
-  map.calendar_feed 'calendar_feeds/grab/:user_id/:token.:format', :controller => 'calendar_feeds', :action => 'grab'
-  map.resources :calendar_feeds
+  match "cas_logout", :controller => 'sessions', :action => 'logout', as: :cas_logout
+  match 'calendar_feeds/grab/:user_id/:token.:format', :controller => 'calendar_feeds', :action => 'grab', as: :calendar_feed
 
   # routes for managing superusers
-  map.superusers "superusers", :controller => 'superusers'
-  map.add_superusers "superusers/add", :controller => 'superusers', :action => 'add'
-  map.remove_superusers "superusers/remove", :controller => 'superusers', :action => 'remove'
 
-  map.resources :user_sessions
+  match "superusers", :controller => 'superusers', action: 'index', as: :superusers
+  match "superusers/add", :controller => 'superusers', :action => 'add', as: :add_superusers
+  match "superusers/remove", :controller => 'superusers', :action => 'remove', as: :remove_superusers
 
-  map.resources :password_resets
-
-  map.resources :user_configs, :only => [:edit, :update]
-
-  map.resources :sub_requests
-  map.resources :notices, :collection => {:archive => :get}
-
-  map.resources :stats,
-              :collection => {:for_user => [:post, :get], :for_location => [:post, :get], :index => [:post, :get]}
-
-  map.resources :payforms,
-              :collection => {:index => [:post, :get], :prune => :delete, :go => :get, :search => [:post, :get]},
-              :member => {:submit => :get, :unsubmit => :get, :approve => :get, :skip => :get, :unskip => :get, :unapprove => :get, :print => :get},
-              :shallow => true do |payform|
-                  payform.resources :payform_items, :member => {:delete => :get}
-              end
-
-  map.resources :payform_item_sets
-  # Temp fix for button not creating POST request for /payform_sets
-  map.create_payform_set '/payform_sets/create', :controller => 'payform_sets', :action => 'create'
-  map.resources :payform_sets
-
-
-
-  map.resources :department_configs, :only => [:edit, :update]
-
-  map.resources :payforms,
-                :collection => { :prune => :delete, :go => :get, :search => :post},
-                :member => {:submit => :get, :unsubmit => :get, :approve => :get, :skip => :get, :unskip => :get, :unapprove => :get, :print => :get},
-                :shallow => true do |payform|
-    payform.resources :payform_items, :member => {:delete => :get}
+  resources :payforms, :shallow => true do
+    collection do
+      get :index
+      post :index
+      delete :prune
+      get :go
+      post :search
+      get :search
+      post :search
+    end
+    member do
+      get :submit
+      get :unsubmit
+      get :approve
+      get :skip
+      get :unskip
+      get :unapprove
+      get :print
+    end
+    resources :payform_items do
+      member do
+        get :delete
+      end
+    end
   end
 
-  map.resources :payform_items
 
-  map.resources :time_slots, :member => {:rerender => :get} #TODO: What should this be nested under, if anything?
+  resources :time_slots do
+    member { get :rerender } #TODO: What should this be nested under, if anything?
+  end
 
-  map.resources :shifts, :new => {:unscheduled => :get, :ajax_create => :post}, :member => {:rerender => :get}, :collection => {:show_active => :get, :show_unscheduled => :get}, :shallow => true do |shifts|
-    shifts.resource :report do |report|
-      report.resources :report_items
+  resources :shifts, :shallow => true do
+    member do
+      get :rerender
+    end
+    collection do
+      get :unscheduled, as: :new_unscheduled
+      get :show_active
+      get :show_unscheduled
+    end
+    resource :report do
+      resources :report_items
     end
     #NOTE: "sub_requests" is a clearer model name, we use subs for routing
-    shifts.resources :sub_requests, :member => {:take => :put, :get_take_info => :get},
-                                    :as => "subs"
+    resources :sub_requests, as: "subs" do
+      member do
+        put :take
+        get :get_take_info
+      end
+    end
   end
 
-
-  map.resources :requested_shifts
-  map.resources :templates
-
-  map.resources :users, :collection => {:update_superusers => :post}, :member => {:toggle => [:get, :post]} do |user|
-    user.resources :punch_clocks
-  end
-  map.resources :reports, :except => [:new], :member => {:popup => :get} do |report|
-    report.resources :report_items
-  end
-
-  map.resources :requested_shifts
-  map.resources :templates
-
-  map.resources :public_view, :member => {:for_location => [:post, :get]}
-
-  map.resources :users, :collection => {:update_superusers => :post}, :member => {:toggle => [:get, :post]} do |user|
-    user.resources :punch_clocks
+  resources :users do
+    collection do
+      post :update_superusers
+    end
+    member do
+      get   :toggle
+      post  :toggle
+    end
+    resources :punch_clocks
   end
 
-  #TODO Fix report items routing, this is temporary
-  map.resources :locations, :except => [:index, :show, :edit, :find_allowed_locations, :new, :update, :create, :destroy], :collection => {:display_report_items => [:get, :post], :toggle => [:get, :post]}
-  #map.resources :locations, :collection => {:display_report_items => [:post, :get], :toggle => [:post, :get], :index => [:post, :get]}, :except => [:index, :show, :edit, :find_allowed_locations, :new, :update, :create, :destroy]
-
-  map.resources :data_types do |data_type|
-    data_type.resources :data_fields
-    data_type.resources :data_objects, :only => [:new, :create, :index]
+  resources :reports, :except => [:new] do
+    member do
+      get :popup
+      get :update_reports
+      post :custom_search
+    end
+    resources :report_items
   end
 
-  map.resources :data_objects do |data_object|
-    data_object.resources :data_entries, :member => {:edit => :put, :update => :put}
+  resources :public_view do
+    member do
+      get  :for_location
+      post :for_location
+    end
   end
 
-  map.resources :departments, :shallow => true do |departments|
-    departments.resources :users, :collection => {:mass_add => :get, :mass_create => :post, :restore => :post, :autocomplete => :get, :search => :post, :import => :get, :save_import => :post}
-    departments.resources :loc_groups
-    departments.resources :locations
-    departments.resources :roles, :member => {:users => [:post,:get] }
-    departments.resources :categories
+  resources :users do
+    collection do
+      post :update_superusers
+    end
+    member do
+      post :toggle
+      get :toggle
+    end
+    resources :punch_clocks
   end
 
-  map.resources :user_profile_fields
-  map.resources :user_profiles, :collection => {:search => :post}
-  map.facebook "/facebook", :controller => "user_profiles", :action => "facebook"
-  map.email_group "/email_group", :controller => "shifts", :action => "email_group"
+  resources :locations, except: [:find_allowed_locations] do
+    collection do
+      get  :display_report_items
+      post :display_report_items
+      get  :toggle
+      post :toggle
+    end
+  end
+  #resources :locations, :collection => {:display_report_items => [:post, :get], :toggle => [:post, :get], :index => [:post, :get]}, :except => [:index, :show, :edit, :find_allowed_locations, :new, :update, :create, :destroy]
 
-  map.resources :punch_clock_sets
+  resources :data_types do
+    resources :data_fields
+    resources :data_objects, only: [:new, :create, :index]
+  end
 
-  # permission is always created indirectly so there is only index method that lists them
-  map.resources :permissions, :only => :index
-  map.resources :stats, :collection => {:for_user => [:post, :get], :for_location => [:post, :get], :index => [:post, :get]}
+  resources :data_objects do
+    resources :data_entries do
+      member do
+        put :edit
+        put :update
+      end
+    end
+  end
+
+  resources :departments, :shallow => true do
+    resources :users do
+      collection do
+        get :import
+        get :mass_add
+        post :mass_create
+        get :autocomplete
+        post :restore
+        post :search
+        post :save_import
+      end
+    end
+    resources :loc_groups do
+      get :toggle, on: :member
+    end
+    resources :locations
+    resources :roles do
+      member do
+        get  :users
+        post :users
+      end
+    end
+    resources :categories
+  end
+
+  resources :user_profile_fields
+  resources :user_profiles do
+    collection do
+      post :search
+    end
+  end
+
+  match "/facebook", :controller => "user_profiles", :action => "facebook", as: :facebook
+  match "/email_group", :controller => "shifts", :action => "email_group", as: :email_group
+
+  resources :punch_clock_sets
+
+  # permissions are always created indirectly so there we only need an index method to list them
+  resources :permissions, :only => :index
+
+  resources :stats do
+    collection do
+      post :for_user
+      get :for_user
+      post :for_location
+      get :for_location
+      get  :index
+      post :index
+    end
+  end
 
   #map.report_items 'report_items/for_location', :controller => 'report_items', :action => 'for_location'
 
-  map.dashboard '/dashboard', :controller => 'dashboard', :action => 'index'
-  map.access_denied '/access_denied', :controller => 'application', :action => 'access_denied'
+  match '/dashboard', :controller => 'dashboard', :action => 'index', as: :dashboard
+  match '/access_denied', :controller => 'application', :action => 'access_denied', as: :access_denied
 
-  map.rt_add_job '/rt', :controller => 'hooks', :action => 'add_job'
-
-	map.resources :templates, :collection => {:update_locations => :post} do |template|
-		template.resources :requested_shifts
-		template.resources :shift_preferences
-		template.resources :template_time_slots
+	resources :templates  do
+    collection do
+      post :update_locations
+    end
+		resources :requested_shifts
+		resources :shift_preferences
+		resources :template_time_slots
   end
 
-  map.root :controller => "dashboard"
+  root to: "dashboard#index"
 
   get '/status' => 'status#index'
 
-  map.connect ':controller/:action/:id'
-  map.connect ':controller/:action/:id.:format'
-
-  map.test_new_lists '/test_new_lists', :controller => 'reports', :action => 'tasks_and_objects_list'
-  map.update_data_objects '/update_data_objects', :controller => 'data_objects', :action => 'update_data_objects'
-  map.active_tasks '/active_tasks', :controller => 'tasks', :action => 'active_tasks'
+  match '/test_new_lists', :controller => 'reports', :action => 'tasks_and_objects_list', as: :test_new_lists
+  match '/update_data_objects', :controller => 'data_objects', :action => 'update_data_objects', as: :update_data_objects
+  match '/active_tasks', :controller => 'tasks', :action => 'active_tasks', as: :active_tasks
 
   match "/delayed_job" => DelayedJobWeb, :anchor => false
 end
