@@ -6,8 +6,9 @@ class Task < ActiveRecord::Base
   validates_uniqueness_of :name
   validates_presence_of :name, :kind
   validate :start_less_than_end
-  validates_format_of :link, with: /^(http|https):\/\/[a-z0-9]+([\-\.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$/ix
-
+  validates :link, allow_blank: true, format: {
+    with: /^(http|https):\/\/[a-z0-9]+([\-\.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$/ix
+  }
 
   scope :active, -> { where(active: true) }
   scope :after_now, -> { where("end >= ?", Time.now.utc) }
