@@ -7,8 +7,8 @@ class UserProfile < ActiveRecord::Base
       styles: { large: "500x500>", medium: "250x250>", small: "150x150>", thumbnail: "100x100#", pico: "22x22#"},
       processors: [:cropper],
       default_url: "/default_attachments/user_profiles/default_:style.jpg",
-      url: "/attachments/user_profiles/:id/:style/:normalized_photo_name.extension",
-      path: ":rails_root/public/attachments/user_profiles/:id/:style/:normalized_photo_name.extension"
+      url: "/attachments/user_profiles/:id/:style/:normalized_photo_name",
+      path: ":rails_root/public/attachments/user_profiles/:id/:style/:normalized_photo_name"
   validates_attachment_size :photo, less_than: 3.megabytes
   validates_attachment_content_type :photo, content_type: ['image/jpeg', 'image/png', 'image/jpg']
 
